+++ Changelog.md
@@ -4,3 +4,5 @@
 - Fix for #1 [2017.03.06]
 - Fix for #2 [2017.03.06]
 - ReadMe updated for Python Script execution [2017.03.06]
+- Support for Whole Show Downloading for Crunchyroll [2017.03.06]
+- Selection of language for the Crunchyroll Show [2017.03.06]